# openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Client Service",
    "description": "User and organization management service for MCPFactory",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "https://client.mcpfactory.org"
    }
  ],
  "components": {
    "securitySchemes": {
      "ApiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "x-api-key"
      }
    },
    "schemas": {
      "HealthResponse": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string"
          },
          "service": {
            "type": "string"
          }
        },
        "required": [
          "status",
          "service"
        ]
      },
      "GetUserResponse": {
        "type": "object",
        "properties": {
          "user": {
            "type": "object",
            "properties": {
              "id": {
                "type": "string",
                "format": "uuid"
              },
              "email": {
                "type": "string",
                "nullable": true
              },
              "firstName": {
                "type": "string",
                "nullable": true
              },
              "lastName": {
                "type": "string",
                "nullable": true
              }
            },
            "required": [
              "id",
              "email",
              "firstName",
              "lastName"
            ]
          }
        },
        "required": [
          "user"
        ]
      },
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string"
          }
        },
        "required": [
          "error"
        ]
      },
      "ListUsersResponse": {
        "type": "object",
        "properties": {
          "users": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ListUsersUser"
            }
          },
          "total": {
            "type": "integer"
          },
          "limit": {
            "type": "integer"
          },
          "offset": {
            "type": "integer"
          }
        },
        "required": [
          "users",
          "total"
        ]
      },
      "ListUsersUser": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "externalId": {
            "type": "string",
            "nullable": true
          },
          "email": {
            "type": "string",
            "nullable": true
          },
          "firstName": {
            "type": "string",
            "nullable": true
          },
          "lastName": {
            "type": "string",
            "nullable": true
          },
          "imageUrl": {
            "type": "string",
            "nullable": true
          },
          "phone": {
            "type": "string",
            "nullable": true
          },
          "createdAt": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "externalId",
          "email",
          "firstName",
          "lastName",
          "imageUrl",
          "phone",
          "createdAt"
        ]
      },
      "PublicStatsResponse": {
        "type": "object",
        "properties": {
          "totalOrgs": {
            "type": "integer"
          },
          "totalUsers": {
            "type": "integer"
          },
          "monthlyGrowth": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/MonthlyGrowthEntry"
            }
          }
        },
        "required": [
          "totalOrgs",
          "totalUsers",
          "monthlyGrowth"
        ]
      },
      "MonthlyGrowthEntry": {
        "type": "object",
        "properties": {
          "month": {
            "type": "string"
          },
          "newOrgs": {
            "type": "integer"
          },
          "newUsers": {
            "type": "integer"
          }
        },
        "required": [
          "month",
          "newOrgs",
          "newUsers"
        ]
      },
      "ResolveResponse": {
        "type": "object",
        "properties": {
          "orgId": {
            "type": "string",
            "format": "uuid"
          },
          "userId": {
            "type": "string",
            "format": "uuid"
          },
          "orgCreated": {
            "type": "boolean"
          },
          "userCreated": {
            "type": "boolean"
          }
        },
        "required": [
          "orgId",
          "userId",
          "orgCreated",
          "userCreated"
        ]
      },
      "ResolveBody": {
        "type": "object",
        "properties": {
          "externalOrgId": {
            "type": "string",
            "minLength": 1
          },
          "externalUserId": {
            "type": "string",
            "minLength": 1
          },
          "email": {
            "type": "string",
            "format": "email"
          },
          "firstName": {
            "type": "string"
          },
          "lastName": {
            "type": "string"
          },
          "imageUrl": {
            "type": "string",
            "format": "uri"
          },
          "orgName": {
            "type": "string"
          },
          "orgSlug": {
            "type": "string",
            "minLength": 1
          }
        },
        "required": [
          "externalOrgId",
          "externalUserId"
        ]
      },
      "ValidateInviteResponse": {
        "type": "object",
        "properties": {
          "valid": {
            "type": "boolean"
          },
          "inviterOrgName": {
            "type": "string"
          }
        },
        "required": [
          "valid"
        ]
      },
      "ValidateInviteBody": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string",
            "minLength": 1
          }
        },
        "required": [
          "code"
        ]
      },
      "ClaimInviteResponse": {
        "type": "object",
        "properties": {
          "ok": {
            "type": "boolean"
          },
          "inviterOrgId": {
            "type": "string",
            "format": "uuid"
          }
        },
        "required": [
          "ok",
          "inviterOrgId"
        ]
      },
      "ClaimInviteCappedResponse": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string"
          },
          "used": {
            "type": "integer"
          },
          "total": {
            "type": "integer"
          }
        },
        "required": [
          "error",
          "used",
          "total"
        ]
      },
      "ClaimInviteBody": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string",
            "minLength": 1
          },
          "inviteeOrgId": {
            "type": "string",
            "format": "uuid"
          }
        },
        "required": [
          "code",
          "inviteeOrgId"
        ]
      },
      "InviteStatusResponse": {
        "type": "object",
        "properties": {
          "used": {
            "type": "integer"
          },
          "total": {
            "type": "integer"
          },
          "code": {
            "type": "string",
            "nullable": true
          },
          "expired": {
            "type": "boolean"
          }
        },
        "required": [
          "used",
          "total",
          "code",
          "expired"
        ]
      },
      "WaitlistRequestResponse": {
        "type": "object",
        "properties": {
          "ok": {
            "type": "boolean",
            "enum": [
              true
            ]
          },
          "position": {
            "type": "integer"
          }
        },
        "required": [
          "ok",
          "position"
        ]
      },
      "WaitlistRequestBody": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "format": "email"
          },
          "brandUrl": {
            "type": "string",
            "minLength": 1
          }
        },
        "required": [
          "email",
          "brandUrl"
        ]
      },
      "WaitlistPositionResponse": {
        "type": "object",
        "properties": {
          "position": {
            "type": "integer"
          }
        },
        "required": [
          "position"
        ]
      }
    },
    "parameters": {}
  },
  "paths": {
    "/health": {
      "get": {
        "summary": "Health check",
        "responses": {
          "200": {
            "description": "Service is healthy",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HealthResponse"
                }
              }
            }
          }
        }
      }
    },
    "/internal/users/{userId}": {
      "get": {
        "summary": "Get a user by internal UUID",
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "required": true,
            "name": "userId",
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "description": "User found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetUserResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "User not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/internal/users": {
      "get": {
        "summary": "List users filtered by org",
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "required": false,
            "name": "orgId",
            "in": "query"
          },
          {
            "schema": {
              "type": "string",
              "minLength": 1
            },
            "required": false,
            "name": "externalOrgId",
            "in": "query"
          },
          {
            "schema": {
              "type": "string"
            },
            "required": false,
            "name": "email",
            "in": "query"
          },
          {
            "schema": {
              "type": "integer",
              "minimum": 1
            },
            "required": false,
            "name": "limit",
            "in": "query"
          },
          {
            "schema": {
              "type": "integer",
              "nullable": true,
              "minimum": 0
            },
            "required": false,
            "name": "offset",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "Users list",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListUsersResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid query parameters",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/internal/orgs/{orgId}/members/{userId}": {
      "get": {
        "summary": "Check if a user is a member of an org",
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "required": true,
            "name": "orgId",
            "in": "path"
          },
          {
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "required": true,
            "name": "userId",
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "description": "User is a member of the org"
          },
          "400": {
            "description": "Invalid parameters",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "User is not a member of the org",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/public/stats/users": {
      "get": {
        "summary": "Get platform-wide stats (total orgs, users, monthly growth)",
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Platform stats",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicStatsResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/internal/resolve": {
      "post": {
        "summary": "Resolve external org/user IDs to internal UUIDs (idempotent upsert)",
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResolveBody"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Identity resolved",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResolveResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request body",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/public/invites/validate": {
      "post": {
        "summary": "Validate an invite code (slug + cap check)",
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ValidateInviteBody"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Validation result (valid=false covers unknown-slug and capped-org)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidateInviteResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request body",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/internal/invites/claim": {
      "post": {
        "summary": "Claim an invite code for a freshly-created org (idempotent)",
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ClaimInviteBody"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Invite claimed (or already claimed by same invitee)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClaimInviteResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request body",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Unknown invite code or invitee org",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "Invite cap reached (3 successful signups already)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClaimInviteCappedResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/internal/orgs/{orgId}/invites/status": {
      "get": {
        "summary": "Get invite usage status for an org",
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "required": true,
            "name": "orgId",
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "description": "Invite status",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InviteStatusResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid orgId",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Org not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/public/waitlist/request-access": {
      "post": {
        "summary": "Request waitlist access (idempotent on email)",
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WaitlistRequestBody"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Waitlist entry created or existing position returned",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WaitlistRequestResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request body",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/public/waitlist/position": {
      "get": {
        "summary": "Get waitlist position for an email",
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "format": "email"
            },
            "required": true,
            "name": "email",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "Position found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WaitlistPositionResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid email",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Email not on waitlist",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    }
  }
}